test(model): unittests for new prediction API
