+++ tests/model/test_train.py
@@ -1,0 +1,5 @@
+from gfibot.model.train import *
+
+
+def test_load_dataset(mock_mongodb):
+    df = load_full_dataset(1)
